# Docs/swagger-Y.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: Handmade
  contact:
    name: Marigrace Silva
    email: mksss161297@gmail.com
host: 127.0.0.1:8000
basePath: /
securityDefinitions: {}
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /handmade/productos/:
    post:
      summary: /handmade/productos/
      tags:
        - Productos
      operationId: /handmade/productos/
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/~1handmade~1productosrequest'
      responses:
        '200':
          description: ''
          headers: {}
    get:
      summary: /handmade/productos/
      tags:
        - Productos
      operationId: /handmade/productos/*
      deprecated: false
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
...